Go toolkit in "create" implemented
Prints info about the toolkits, asks for which toolkit to use.

If Go toolkit selected it also asks for the package ID/name.

If Go toolkit : the step is generated into GOPATH, instead of just a relative dir.

Evaluate minimal revision, to make it possible to unit test it + a step.yml template unit test added.

generateIDFromString (which generates a step id) now keeps numbers instead of replacing them with `-`.